feat(space-onboarding): move a theme manager outside of space (so we can load a default theme preference before we load a space); rename onboarding components (use wizard so it's clear how they suppose to work)
